tuna: Revise the configuration API
This change brings the MaguroHTTP configuration API more in line with 
other application configurations. 

With this change the parameters are more understandable and more fault 
tolerant. 

This change breaks API compatibility with r6 and lower. Considering that 
r7 will bring more API breakage, this was the right time to overhaul the 
configuration API.